refactor(agent): generic chat pre-handler extension point; move Binance dispatch to a plugin (arch #12092 item 1)
The host chat loop hardcoded a 1,102-line Binance DeFi skill dispatch
(`binance-skill-helpers.ts`) that short-circuited generation before normal
action processing, so every chat request on every platform paid for
Binance-specific trigger parsing.

Replace it with a generic, vendor-neutral extension point:

- core: `ChatPreHandler` contract (`types/chat-pre-handler.ts`) +
  `ChatPreHandlerRegistry` (`runtime/chat-pre-handler-registry.ts`).
  `Plugin.chatPreHandlers` registers handlers; the runtime exposes
  `registerChatPreHandler(s)` / `drainChatPreHandlers(ctx)`. The registry
  drains handlers by descending priority; first non-null result wins.
- agent: the chat loop calls only `runtime.drainChatPreHandlers(...)`.
  The generic fallback-action helpers (`executeFallbackParsedActions`,
  `parseFallbackActionBlocks`) move to `fallback-action-helpers.ts`; the
  Binance file and its short-circuit are deleted. No Binance identifiers
  remain in `packages/agent/src` (only unrelated BSC RPC hostnames).
- plugin-agent-skills: the Binance trigger regexes + dispatch move to
  `src/binance/` and self-register as a chat pre-handler (priority 100),
  preserving identical dispatch behavior for the existing skill triggers.

Behavior-preserving: triggers dispatch identically via the registered
pre-handler; non-matching messages fall through to normal processing.
